requirements.txt
fastapi == 0.95.0
uvicorn[standard]==0.21.1
pydantic == 1.10.7
tortoise-orm[asyncpg] == 0.19.3
pyjwt == 2.6.0